기능 완성 / 게시판 조회수
쿠키에 따라서 오전 12시마다 초기화 되도록 설정했음.
쿠키에 본 게시물들 일련번호를 담아놓고 일치하면 조회수 안 증가하게 해놓게 함. 단점이 있다면 길이가 너무 길어져서 이 기능 말고도 다른 점을 구상해봐야 할듯

+++ src/main/java/com/example/teto/board/entity/Board.java
@@ -32,10 +32,10 @@ public class Board {
     private Date date;
 
     @Column(name = "views", nullable = false)
-    private Integer views = 0;
+    private Long views = 0L;
 
     @Column(name = "likes", nullable = false)
-    private Integer likes = 0;
+    private Long likes = 0L;
 
     @OneToMany(mappedBy = "likeBoard")
     private List<LikeBoards> likeBoards = new ArrayList<>();
